Kall aktørregister gjennom proxy
