+++ Dockerfile
@@ -1,6 +1,8 @@
 FROM navikt/node-express:1.0.0
 WORKDIR /app
 
+RUN yarn add http-proxy-middleware
+
 COPY build/ build/
 COPY src/server/ src/server/
