+++ package.json
@@ -68,7 +68,8 @@
         "axios-mock-adapter": "^1.16.0",
         "body-parser": "^1.18.3",
         "express": "^4.16.4",
-        "prettier": "^1.17.1"
+        "prettier": "^1.17.1",
+        "http-proxy-middleware": "^0.19.1"
     },
     "scripts": {
         "start": "craco start",
